pyproject.toml
[tool.poetry]
name = "de_stock_ingestion"
version = "0.1.0"
description = "End-to-end data engineering project with Medallion Architecture"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.9"
requests = "^2.31.0"
# boto3 = "^1.34.0"  # Commented out AWS SDK
google-cloud-storage = "^2.13.0"  # Added Google Cloud Storage
google-cloud-bigquery = "^3.17.0"  # Added BigQuery client
pandas = "^2.1.0"
yfinance = "^0.2.36"
python-dotenv = "^1.0.0"
structlog = "^23.2.0"
numpy = "<2.0.0"
jsonschema = "^4.21.1"  # Added for JSON schema validation

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-mock = "^3.12.0"
black = "^23.7.0"
isort = "^5.12.0"
flake8 = "^6.1.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py39']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3 